Test rv32ui-p-addi  passed

+++ src/cpu/decode/mmap_parse_16.rs
@@ -1,4 +1,5 @@
-use super::{OpecodeKind, Instruction, Decode};
+use super::Decode;
+use crate::cpu::instruction::{OpecodeKind, Instruction};
 
 fn quadrant0(opmap: &u8) -> Result<OpecodeKind, &'static str> {
     match opmap {
@@ -87,10 +88,10 @@ impl Decode for u16 {
             Ok(opc)  => opc,
             Err(msg) => panic!("{}, {:b}", msg, self),
         };
-        let new_rd:  Option<u8>  = self.parse_rd(&new_opc);
-        let new_rs1: Option<u8>  = self.parse_rs1(&new_opc);
-        let new_rs2: Option<u8>  = self.parse_rs2(&new_opc);
-        let new_imm: Option<u32> = self.parse_imm(&new_opc);
+        let new_rd:  Option<usize>  = self.parse_rd(&new_opc);
+        let new_rs1: Option<usize>  = self.parse_rs1(&new_opc);
+        let new_rs2: Option<usize>  = self.parse_rs2(&new_opc);
+        let new_imm: Option<i32> = self.parse_imm(&new_opc);
 
         Instruction {
             opc: new_opc,
@@ -115,11 +116,11 @@ impl Decode for u16 {
         }
     }
 
-    fn parse_rd(&self, opkind: &OpecodeKind) -> Option<u8> {
+    fn parse_rd(&self, opkind: &OpecodeKind) -> Option<usize> {
         let inst: &u16 = self;
-        let q0_rd: u8  = ((inst >> 2) & 0x7) as u8;
-        let q1_rd: u8  = ((inst >> 7) & 0x7) as u8;
-        let q2_rd: u8  = ((inst >> 7) & 0x1F) as u8;
+        let q0_rd: usize  = ((inst >> 2) & 0x7) as usize;
+        let q1_rd: usize  = ((inst >> 7) & 0x7) as usize;
+        let q2_rd: usize  = ((inst >> 7) & 0x1F) as usize;
 
         match opkind {
             // Quadrant 0
@@ -128,33 +129,33 @@ impl Decode for u16 {
             OpecodeKind::OP_C_LW        => Some(q0_rd),
             OpecodeKind::OP_C_FLW       => Some(q0_rd),
             // Quadrant 1
-            OpecodeKind::OP_C_SRLI        => Some(q1_rd),
-            OpecodeKind::OP_C_SRAI        => Some(q1_rd),
-            OpecodeKind::OP_C_ANDI        => Some(q1_rd),
-            OpecodeKind::OP_C_SUB        => Some(q1_rd),
-            OpecodeKind::OP_C_XOR        => Some(q1_rd),
+            OpecodeKind::OP_C_SRLI      => Some(q1_rd),
+            OpecodeKind::OP_C_SRAI      => Some(q1_rd),
+            OpecodeKind::OP_C_ANDI      => Some(q1_rd),
+            OpecodeKind::OP_C_SUB       => Some(q1_rd),
+            OpecodeKind::OP_C_XOR       => Some(q1_rd),
             OpecodeKind::OP_C_OR        => Some(q1_rd),
-            OpecodeKind::OP_C_AND        => Some(q1_rd),
+            OpecodeKind::OP_C_AND       => Some(q1_rd),
             // Quadrant 2
-            OpecodeKind::OP_C_SLLI        => Some(q2_rd),
-            OpecodeKind::OP_C_FLDSP        => Some(q2_rd),
-            OpecodeKind::OP_C_LWSP        => Some(q2_rd),
-            OpecodeKind::OP_C_FLWSP        => Some(q2_rd),
+            OpecodeKind::OP_C_SLLI      => Some(q2_rd),
+            OpecodeKind::OP_C_FLDSP     => Some(q2_rd),
+            OpecodeKind::OP_C_LWSP      => Some(q2_rd),
+            OpecodeKind::OP_C_FLWSP     => Some(q2_rd),
             OpecodeKind::OP_C_JR        => Some(q2_rd),
             OpecodeKind::OP_C_MV        => Some(q2_rd),
-            OpecodeKind::OP_C_EBREAK       => Some(q2_rd),
-            OpecodeKind::OP_C_JALR        => Some(q2_rd),
-            OpecodeKind::OP_C_ADD        => Some(q2_rd),
+            OpecodeKind::OP_C_EBREAK    => Some(q2_rd),
+            OpecodeKind::OP_C_JALR      => Some(q2_rd),
+            OpecodeKind::OP_C_ADD       => Some(q2_rd),
             _ => None,
         }
     }
 
-    fn parse_rs1(&self, opkind: &OpecodeKind) -> Option<u8> {
+    fn parse_rs1(&self, opkind: &OpecodeKind) -> Option<usize> {
         let inst: &u16 = self;
-        let q0_rs1: u8 = ((inst >> 7) & 0x3) as u8;
-        let q1_rs1: u8 = ((inst >> 7) & 0x3) as u8;
-        let q2_rs1: u8 = ((inst >> 7) & 0x3) as u8;
-        let addi_rs1: u8 = ((inst >> 7) & 0x1F) as u8;
+        let q0_rs1: usize = ((inst >> 7) & 0x3) as usize;
+        let q1_rs1: usize = ((inst >> 7) & 0x3) as usize;
+        let q2_rs1: usize = ((inst >> 7) & 0x3) as usize;
+        let addi_rs1: usize = ((inst >> 7) & 0x1F) as usize;
 
         match opkind {
             // Quadrant 0
@@ -185,11 +186,11 @@ impl Decode for u16 {
         }
     }
 
-    fn parse_rs2(&self, opkind: &OpecodeKind) -> Option<u8> {
+    fn parse_rs2(&self, opkind: &OpecodeKind) -> Option<usize> {
         let inst: &u16 = self;
-        let q0_rs2: u8 = ((inst >> 2) & 0x7) as u8;
-        let q1_rs2: u8 = ((inst >> 2) & 0x7) as u8;
-        let q2_rs2: u8 = ((inst >> 2) & 0x1F) as u8;
+        let q0_rs2: usize = ((inst >> 2) & 0x7) as usize;
+        let q1_rs2: usize = ((inst >> 2) & 0x7) as usize;
+        let q2_rs2: usize = ((inst >> 2) & 0x1F) as usize;
 
         match opkind {
             // Quadrant 0
@@ -211,13 +212,13 @@ impl Decode for u16 {
         }
     }
 
-    fn parse_imm(&self, opkind: &OpecodeKind) -> Option<u32> {
-        let q0_imm = ((self >> 5) & 0x3 + (((self >> 10) & 0x7) << 0x5)) as u32;
-        let q1_imm = ((self >> 2) & 0x1F + (((self >> 12) & 0x1) << 0x2)) as u32;
-        let q2_imm = ((self >> 2) & 0x1F + (((self >> 12) & 0x1) << 0x2)) as u32;
+    fn parse_imm(&self, opkind: &OpecodeKind) -> Option<i32> {
+        let q0_imm = ((self >> 5) & 0x3 + (((self >> 10) & 0x7) << 0x5)) as i32;
+        let q1_imm = ((self >> 2) & 0x1F + (((self >> 12) & 0x1) << 0x2)) as i32;
+        let q2_imm = ((self >> 2) & 0x1F + (((self >> 12) & 0x1) << 0x2)) as i32;
         match opkind {
             // Quadrant0
-            OpecodeKind::OP_C_ADDI4SPN  => Some(((self >> 5) & 0xFF) as u32),
+            OpecodeKind::OP_C_ADDI4SPN  => Some(((self >> 5) & 0xFF) as i32),
             OpecodeKind::OP_C_FLD       => Some(q0_imm),
             OpecodeKind::OP_C_LW        => Some(q0_imm),
             OpecodeKind::OP_C_FLW       => Some(q0_imm),
@@ -227,27 +228,26 @@ impl Decode for u16 {
             // Quadrant1
             OpecodeKind::OP_C_NOP       => Some(q1_imm),
             OpecodeKind::OP_C_ADDI      => Some(q1_imm),
-            OpecodeKind::OP_C_JAL       => Some(((self >> 2) & 0x7FF) as u32),
+            OpecodeKind::OP_C_JAL       => Some(((self >> 2) & 0x7FF) as i32),
             OpecodeKind::OP_C_LI        => Some(q1_imm),
             OpecodeKind::OP_C_ADDI16SP  => Some(q1_imm),
             OpecodeKind::OP_C_LUI       => Some(q1_imm),
             OpecodeKind::OP_C_SRLI      => Some(q1_imm),
             OpecodeKind::OP_C_SRAI      => Some(q1_imm),
             OpecodeKind::OP_C_ANDI      => Some(q1_imm),
-            OpecodeKind::OP_C_J         => Some(((self >> 2) & 0x7FF) as u32),
-            OpecodeKind::OP_C_BEQZ      => Some(((self >> 2) & 0x1F + (((self >> 10) & 0x7) << 0x2)) as u32),
-            OpecodeKind::OP_C_BNEZ      => Some(((self >> 2) & 0x1F + (((self >> 10) & 0x7) << 0x2)) as u32),
+            OpecodeKind::OP_C_J         => Some(((self >> 2) & 0x7FF) as i32),
+            OpecodeKind::OP_C_BEQZ      => Some(((self >> 2) & 0x1F + (((self >> 10) & 0x7) << 0x2)) as i32),
+            OpecodeKind::OP_C_BNEZ      => Some(((self >> 2) & 0x1F + (((self >> 10) & 0x7) << 0x2)) as i32),
             // Quadrant2
             OpecodeKind::OP_C_SLLI      => Some(q2_imm),
             OpecodeKind::OP_C_FLDSP     => Some(q2_imm),
             OpecodeKind::OP_C_LWSP      => Some(q2_imm),
             OpecodeKind::OP_C_FLWSP     => Some(q2_imm),
             OpecodeKind::OP_C_JR        => Some(q2_imm),
-            OpecodeKind::OP_C_FSDSP     => Some(((self >> 7) & 0x3F) as u32),
-            OpecodeKind::OP_C_SWSP      => Some(((self >> 7) & 0x3F) as u32),
-            OpecodeKind::OP_C_FSWSP     => Some(((self >> 7) & 0x3F) as u32),
+            OpecodeKind::OP_C_FSDSP     => Some(((self >> 7) & 0x3F) as i32),
+            OpecodeKind::OP_C_SWSP      => Some(((self >> 7) & 0x3F) as i32),
+            OpecodeKind::OP_C_FSWSP     => Some(((self >> 7) & 0x3F) as i32),
             _ => None,
         }
     }
 }
-
